[server][test] Fix a build error.
DBAgentTest.cc:681:2: error: call of overloaded 'assert_equal(const size_t&, uint64_t, const char [14], const char [34])' is ambiguous
DBAgentTest.cc:681:2: note: candidates are:
In file included from /usr/local/include/cutter/cppcutter/cppcut-assertions.h:23:0,
                 from /usr/local/include/cutter/cppcutter.h:25,
                 from DBAgentTest.h:23,
                 from DBAgentTest.cc:20:
/usr/local/include/cutter/cppcutter/cppcut-assertions-helper.h:54:10: note: void cut::assert_equal(char*, char*, const char*, const char*) <near match>
/usr/local/include/cutter/cppcutter/cppcut-assertions-helper.h:54:10: note:   no known conversion for argument 2 from 'uint64_t {aka long long unsigned int}' to 'char*'
/usr/local/include/cutter/cppcutter/cppcut-assertions-helper.h:58:10: note: void cut::assert_equal(const char*, const char*, const char*, const char*) <near match>
/usr/local/include/cutter/cppcutter/cppcut-assertions-helper.h:58:10: note:   no known conversion for argument 2 from 'uint64_t {aka long long unsigned int}' to 'const char*'

+++ server/test/DBAgentTest.cc
@@ -678,7 +678,8 @@ void dbAgentGetNumberOfAffectedRows(DBAgent &dbAgent, DBAgentChecker &checker)
 	DBAgentDeleteArg arg;
 	arg.tableName = TABLE_NAME_TEST;
 	dbAgent.deleteRows(arg);
-	cppcut_assert_equal(NUM_TEST_DATA, dbAgent.getNumberOfAffectedRows());
+	cppcut_assert_equal(static_cast<uint64_t>(NUM_TEST_DATA),
+			    dbAgent.getNumberOfAffectedRows());
 }
 
 // --------------------------------------------------------------------------
